+++ CHANGELOG.md
@@ -24,7 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * clippy: `redundant_static_lifetimes`, `needless_borrow`, `bool_assert_comparison`
 * rust-version: "1.56.0" to "1.60.0"
 * bug: file compare on windows
-
+* bug: on macos:
+  - zcat: can't stat: fixtures/text10k.text.gz (fixtures/text10k.text.gz.Z): No such file or directory
+  must be used `gzcat` on macos
 
 ## [0.1.24] (2023-01-11)
 ### Added